orginazed imports and fixed other issues
